src/requirements.txt
Django==4.2.28
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
cloudinary==1.44.1
django-cloudinary-storage==0.3.0
python-decouple==3.8
requests==2.32.5
Pillow
drf-spectacular
<<<<<<< HEAD:requirements.txt
psycopg2-binary==2.9.10
gunicorn==25.1.0
whitenoise==6.12.0
=======
gunicorn
>>>>>>> origin/main:src/requirements.txt
